Make the config event publish an ACK event + make it more resilient to input errors

+++ src/nostr/cardConfig.ts
@@ -1,6 +1,6 @@
 import type { NDKFilter, NostrEvent } from '@nostr-dev-kit/ndk';
 
-import { Kind } from '@lib/event';
+import { Kind, buildMultiNip04Event } from '@lib/event';
 import { nowInSeconds, requiredEnvVar } from '@lib/utils';
 
 import {
@@ -9,11 +9,12 @@ import {
   Limit,
   parseCardConfigEvent,
   CardStatus,
+  buildCardConfigPayload,
 } from '@lib/config';
 
 import { Context } from '@type/request';
 
-import { Prisma } from '@prisma/client';
+import { Prisma, PrismaClient } from '@prisma/client';
 
 const filter: NDKFilter = {
   kinds: [Kind.PARAMETRIZED_REPLACEABLE.valueOf()],
@@ -63,34 +64,36 @@ const getHandler = (ctx: Context): ((event: NostrEvent) => void) => {
     );
 
     await ctx.prisma.$transaction(async (tx) => {
-      const merchantPubKeys: string[] = (
-        await tx.merchant.findMany({
-          select: {
-            pubKey: true,
-          },
-        })
-      ).map((merchant: { pubKey: string }): string => merchant.pubKey);
-      await tx.trustedMerchants.deleteMany({
-        where: {
-          holderPubKey: holderPubKey,
-        },
-      });
-      await tx.trustedMerchants.createMany({
-        data: content['trusted-merchants']
-          .filter((trustedMerchant: { pubkey: string }): boolean =>
-            merchantPubKeys.includes(trustedMerchant.pubkey),
-          )
-          .map(
-            (trustedMerchant: {
-              pubkey: string;
-            }): Prisma.TrustedMerchantsCreateManyInput => {
-              return {
-                holderPubKey: holderPubKey,
-                merchantPubKey: trustedMerchant.pubkey,
-              };
+      if ('trusted-merchants' in content) {
+        const merchantPubKeys: string[] = (
+          await tx.merchant.findMany({
+            select: {
+              pubKey: true,
             },
-          ),
-      });
+          })
+        ).map((merchant: { pubKey: string }): string => merchant.pubKey);
+        await tx.trustedMerchants.deleteMany({
+          where: {
+            holderPubKey: holderPubKey,
+          },
+        });
+        await tx.trustedMerchants.createMany({
+          data: content['trusted-merchants']
+            .filter((trustedMerchant: { pubkey: string }): boolean =>
+              merchantPubKeys.includes(trustedMerchant.pubkey),
+            )
+            .map(
+              (trustedMerchant: {
+                pubkey: string;
+              }): Prisma.TrustedMerchantsCreateManyInput => {
+                return {
+                  holderPubKey: holderPubKey,
+                  merchantPubKey: trustedMerchant.pubkey,
+                };
+              },
+            ),
+        });
+      }
 
       const holderCardUuids: string[] = (
         await tx.card.findMany({
@@ -103,47 +106,69 @@ const getHandler = (ctx: Context): ((event: NostrEvent) => void) => {
         })
       ).map((x: { uuid: string }) => x.uuid);
 
-      const cardUuids: string[] = Object.keys(content.cards).filter(
-        (uuid: string): boolean => holderCardUuids.includes(uuid),
-      );
+      const modifications: Promise<unknown>[] = [];
 
-      await tx.limit.deleteMany({
-        where: {
-          cardUuid: {
-            in: cardUuids,
-          },
-        },
-      });
-      await tx.limit.createMany({
-        data: ([] as Prisma.LimitCreateManyInput[]).concat(
-          ...cardUuids.map((uuid: string): Prisma.LimitCreateManyInput[] => {
-            return content.cards[uuid].limits.map(
-              (limit: Limit): Prisma.LimitCreateManyInput => {
-                return {
-                  cardUuid: uuid,
-                  amount: limit.amount,
-                  delta: limit.delta,
-                  description: limit.description,
-                  name: limit.name,
-                  token: limit.token,
-                };
-              },
-            );
-          }),
+      for (const uuid in content.cards) {
+        if (!holderCardUuids.includes(uuid)) {
+          continue;
+        }
+        if ('limits' in content.cards[uuid]) {
+          modifications.push(
+            (async function (cardUuid: string) {
+              await tx.limit.deleteMany({ where: { cardUuid } });
+              await tx.limit.createMany({
+                data: content.cards[uuid].limits.map(
+                  (limit: Limit): Prisma.LimitCreateManyInput => {
+                    return {
+                      cardUuid,
+                      amount: limit.amount,
+                      delta: limit.delta,
+                      description: limit.description,
+                      name: limit.name,
+                      token: limit.token,
+                    };
+                  },
+                ),
+              });
+            })(uuid),
+          );
+        }
+
+        const cardUpdate: Prisma.CardUpdateInput = {};
+        if ('name' in content.cards[uuid]) {
+          cardUpdate.name = content.cards[uuid].name;
+        }
+        if ('description' in content.cards[uuid]) {
+          cardUpdate.description = content.cards[uuid].description;
+        }
+        if ('status' in content.cards[uuid]) {
+          cardUpdate.enabled =
+            content.cards[uuid].status === CardStatus.ENABLED;
+        }
+
+        modifications.push(
+          tx.card.update({ where: { uuid }, data: cardUpdate }),
+        );
+      }
+
+      await Promise.all(modifications);
+
+      const configAckEvent: NostrEvent = await buildMultiNip04Event(
+        JSON.stringify(
+          await buildCardConfigPayload(holderPubKey, tx as PrismaClient),
         ),
-      });
-      await Promise.all(
-        cardUuids.map((uuid) =>
-          tx.card.update({
-            where: { uuid },
-            data: {
-              description: content.cards[uuid].description,
-              enabled: content.cards[uuid].status === CardStatus.ENABLED,
-              name: content.cards[uuid].name,
-            },
-          }),
-        ),
+        cardPrivateKey,
+        cardPublicKey,
+        [cardPublicKey, holderPubKey],
       );
+      configAckEvent.kind = Kind.PARAMETRIZED_REPLACEABLE.valueOf();
+      configAckEvent.tags.concat([
+        ['e', event.id!],
+        ['t', ConfigTypes.CONFIG.valueOf()],
+        ['d', `${holderPubKey}:${ConfigTypes.CONFIG.valueOf()}`],
+      ]);
+
+      ctx.outbox.publish(configAckEvent);
     });
   };
 };
